log ingest wip
