+++ cmd/ingest_metrics_test.go
@@ -15,98 +15,101 @@
 package cmd
 
 import (
+	"log/slog"
 	"testing"
 
 	"github.com/stretchr/testify/require"
+
+	"github.com/cardinalhq/lakerunner/internal/filereader"
 )
 
-func TestHandleHistogram(t *testing.T) {
-	type testCase struct {
-		name         string
-		bucketCounts []float64
-		bucketBounds []float64
-		wantCounts   []float64
-		wantValues   []float64
+// TestHandleHistogram was moved to the proto reader - no longer needed here
+// The proto reader now handles all histogram processing internally
+
+func TestMetricWriterManager(t *testing.T) {
+	tmpdir := t.TempDir()
+	orgID := "test-org"
+	ingestDateint := int32(20250101)
+	rpfEstimate := int64(1000)
+	ll := slog.Default()
+
+	wm := newMetricWriterManager(tmpdir, orgID, ingestDateint, rpfEstimate, ll)
+	require.NotNil(t, wm)
+	require.Equal(t, tmpdir, wm.tmpdir)
+	require.Equal(t, orgID, wm.orgID)
+	require.Equal(t, ingestDateint, wm.ingestDateint)
+	require.Equal(t, rpfEstimate, wm.rpfEstimate)
+	require.NotNil(t, wm.writers)
+	require.Equal(t, ll, wm.ll)
+}
+
+func TestMetricWriterManager_ProcessRow(t *testing.T) {
+	tmpdir := t.TempDir()
+	wm := newMetricWriterManager(tmpdir, "test-org", int32(20250101), int64(1000), slog.Default())
+
+	// Test metric row
+	row := filereader.Row{
+		"_cardinalhq.timestamp":   int64(1640995200000),
+		"_cardinalhq.metric_type": "gauge",
+		"_cardinalhq.name":        "cpu.usage",
+		"_cardinalhq.value":       float64(75.5),
+		"host":                    "web-server-1",
+		"region":                  "us-west-2",
 	}
 
-	tests := []testCase{
-		{
-			name:         "simple one bucket",
-			bucketCounts: []float64{2},
-			bucketBounds: []float64{10},
-			wantCounts:   []float64{2},
-			wantValues:   []float64{5}, // (1e-10 + 10)/2
-		},
-		{
-			name:         "two buckets, only second has count",
-			bucketCounts: []float64{0, 3},
-			bucketBounds: []float64{10, 20},
-			wantCounts:   []float64{3},
-			wantValues:   []float64{15}, // (10 + 20)/2
-		},
-		{
-			name:         "three buckets, last is overflow",
-			bucketCounts: []float64{0, 0, 5},
-			bucketBounds: []float64{10, 20},
-			wantCounts:   []float64{5},
-			wantValues:   []float64{21}, // min(20+1, maxTrackableValue)
-		},
-		{
-			name:         "all buckets have counts",
-			bucketCounts: []float64{1, 2, 3},
-			bucketBounds: []float64{10, 20},
-			wantCounts:   []float64{1, 2, 3},
-			wantValues:   []float64{5, 15, 21},
-		},
-		{
-			name:         "empty input",
-			bucketCounts: []float64{},
-			bucketBounds: []float64{},
-			wantCounts:   []float64{},
-			wantValues:   []float64{},
-		},
-		{
-			name:         "overflow bucket with huge bound",
-			bucketCounts: []float64{0, 0, 1},
-			bucketBounds: []float64{10, 1e10},
-			wantCounts:   []float64{1},
-			wantValues:   []float64{1e9}, // capped at maxTrackableValue
-		},
-		{
-			name:         "live data #1",
-			bucketCounts: []float64{0, 409, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-			bucketBounds: []float64{0, 5, 10, 25, 50, 75, 100, 250, 500, 750, 1000, 2500, 5000, 7500, 10000},
-			wantCounts:   []float64{409},
-			wantValues:   []float64{2.5},
-		},
-		{
-			name:         "live data #2",
-			bucketCounts: []float64{0, 299, 60, 7, 1, 2, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0},
-			bucketBounds: []float64{0, 5, 10, 25, 50, 75, 100, 250, 500, 750, 1000, 2500, 5000, 7500, 10000},
-			wantCounts:   []float64{299, 60, 7, 1, 2, 1, 1},
-			wantValues: []float64{
-				(0.0 + 5.0) / 2,
-				(5.0 + 10.0) / 2,
-				(10.0 + 25.0) / 2,
-				(25.0 + 50.0) / 2,
-				(50.0 + 75.0) / 2,
-				(75.0 + 100.0) / 2,
-				(100.0 + 250.0) / 2,
-			},
-		},
+	err := wm.processRow(row)
+	require.NoError(t, err)
+}
+
+func TestMetricWriterManager_ProcessMultipleValues(t *testing.T) {
+	tmpdir := t.TempDir()
+	wm := newMetricWriterManager(tmpdir, "test-org", int32(20250101), int64(1000), slog.Default())
+
+	// Test multiple metric rows
+	row1 := filereader.Row{
+		"_cardinalhq.timestamp":   int64(1640995200000),
+		"_cardinalhq.metric_type": "gauge",
+		"_cardinalhq.name":        "cpu.usage",
+		"_cardinalhq.value":       float64(75.5),
+		"host":                    "web-server-1",
+		"region":                  "us-west-2",
 	}
 
-	const delta = 1e-8
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			gotCounts, gotValues := handleHistogram(tc.bucketCounts, tc.bucketBounds)
-			require.Equal(t, len(gotCounts), len(gotValues), "counts and values length mismatch")
-			require.Equal(t, tc.wantCounts, gotCounts, "counts mismatch")
-			require.Equal(t, len(tc.wantValues), len(gotValues), "values length mismatch", gotValues)
-			for i := range tc.wantCounts {
-				require.InDelta(t, tc.wantValues[i], gotValues[i], delta, "values[%d]", i)
-			}
-		})
+	row2 := filereader.Row{
+		"_cardinalhq.timestamp":   int64(1640995200000),
+		"_cardinalhq.metric_type": "gauge",
+		"_cardinalhq.name":        "cpu.usage",
+		"_cardinalhq.value":       float64(82.3),
+		"host":                    "web-server-1",
+		"region":                  "us-west-2",
 	}
+
+	err := wm.processRow(row1)
+	require.NoError(t, err)
+
+	err = wm.processRow(row2)
+	require.NoError(t, err)
+}
+
+func TestMetricTranslator(t *testing.T) {
+	translator := &MetricTranslator{
+		orgID:    "test-org",
+		bucket:   "test-bucket",
+		objectID: "metrics/test.json.gz",
+	}
+
+	row := filereader.Row{
+		"_cardinalhq.name": "cpu.usage",
+		"host":             "web-server-1",
+	}
+
+	err := translator.TranslateRow(&row)
+	require.NoError(t, err)
+
+	require.Equal(t, "test-bucket", row["resource.bucket.name"])
+	require.Equal(t, "./metrics/test.json.gz", row["resource.file.name"])
+	require.Equal(t, "test-org", row["_cardinalhq.customer_id"])
+	require.Equal(t, "metrics", row["_cardinalhq.telemetry_type"])
+	require.Equal(t, "cpu.usage", row["_cardinalhq.name"]) // Original field preserved
+	require.Equal(t, "web-server-1", row["host"])          // Original field preserved
 }
