+++ internal/parquetwriter/factories/metrics.go
@@ -18,24 +18,37 @@ import (
 	"fmt"
 	"slices"
 
+	mapset "github.com/deckarep/golang-set/v2"
+
+	"github.com/cardinalhq/lakerunner/internal/fingerprint"
 	"github.com/cardinalhq/lakerunner/internal/parquetwriter"
 )
 
 // NewMetricsWriter creates a writer optimized for metrics data.
-// Metrics are expected to be pre-ordered by TID and should not be split across TID boundaries.
+// Metrics are ordered by metric name and grouped for efficient fingerprinting.
 func NewMetricsWriter(baseName, tmpdir string, targetFileSize int64, bytesPerRecord float64) (*parquetwriter.UnifiedWriter, error) {
 	config := parquetwriter.WriterConfig{
 		BaseName:       baseName,
 		TmpDir:         tmpdir,
 		TargetFileSize: targetFileSize,
 
-		// Metrics are already TID-ordered from the TID merger
-		OrderBy: parquetwriter.OrderPresumed,
+		// Order by [metric name, TID] for efficient grouping
+		OrderBy: parquetwriter.OrderMergeSort,
+		OrderKeyFunc: func(row map[string]any) any {
+			name, nameOk := row["_cardinalhq.name"].(string)
+			tid, tidOk := row["_cardinalhq.tid"].(int64)
+			if nameOk && tidOk {
+				return fmt.Sprintf("%s:%d", name, tid)
+			}
+			return ""
+		},
 
-		// Group by TID and never split TID groups across files
+		// Group by [metric name, TID] - don't split groups with same name+TID
 		GroupKeyFunc: func(row map[string]any) any {
-			if tid, ok := row["_cardinalhq.tid"].(int64); ok {
-				return tid
+			name, nameOk := row["_cardinalhq.name"].(string)
+			tid, tidOk := row["_cardinalhq.tid"].(int64)
+			if nameOk && tidOk {
+				return fmt.Sprintf("%s:%d", name, tid)
 			}
 			return nil
 		},
@@ -48,68 +61,76 @@ func NewMetricsWriter(baseName, tmpdir string, targetFileSize int64, bytesPerRec
 	return parquetwriter.NewUnifiedWriter(config)
 }
 
-// MetricsStatsProvider collects TID-related statistics for metrics files.
+// MetricsStatsProvider collects metric name statistics for metrics files.
 type MetricsStatsProvider struct{}
 
 func (p *MetricsStatsProvider) NewAccumulator() parquetwriter.StatsAccumulator {
 	return &MetricsStatsAccumulator{
-		tidSeen: make(map[int64]bool),
+		metricNames: mapset.NewSet[string](),
 	}
 }
 
 // MetricsStatsAccumulator collects metrics-specific statistics.
 type MetricsStatsAccumulator struct {
-	tidSeen map[int64]bool
-	minTID  int64
-	maxTID  int64
-	first   bool
+	metricNames mapset.Set[string]
+	firstTS     int64
+	lastTS      int64
+	first       bool
 }
 
 func (a *MetricsStatsAccumulator) Add(row map[string]any) {
-	if tid, ok := row["_cardinalhq.tid"].(int64); ok {
-		a.tidSeen[tid] = true
+	// Track metric name for fingerprinting
+	if name, ok := row["_cardinalhq.name"].(string); ok && name != "" {
+		a.metricNames.Add(name)
+	}
 
+	// Track timestamp range
+	if ts, ok := row["_cardinalhq.timestamp"].(int64); ok {
 		if !a.first {
-			a.minTID = tid
-			a.maxTID = tid
+			a.firstTS = ts
+			a.lastTS = ts
 			a.first = true
 		} else {
-			if tid < a.minTID {
-				a.minTID = tid
+			if ts < a.firstTS {
+				a.firstTS = ts
 			}
-			if tid > a.maxTID {
-				a.maxTID = tid
+			if ts > a.lastTS {
+				a.lastTS = ts
 			}
 		}
 	}
 }
 
 func (a *MetricsStatsAccumulator) Finalize() any {
-	// Extract sorted list of TIDs from the map
-	tids := make([]int64, 0, len(a.tidSeen))
-	for tid := range a.tidSeen {
-		tids = append(tids, tid)
+	// Create a map with metric names as a set for fingerprinting
+	tagValuesByName := map[string]mapset.Set[string]{
+		"_cardinalhq.name": a.metricNames,
 	}
-	slices.Sort(tids)
+
+	// Generate fingerprints using the same approach as buffet
+	fingerprintSet := fingerprint.ToFingerprints(tagValuesByName)
+	fingerprints := fingerprintSet.ToSlice()
+	slices.Sort(fingerprints)
 
 	return MetricsFileStats{
-		TIDCount: int32(len(a.tidSeen)),
-		MinTID:   a.minTID,
-		MaxTID:   a.maxTID,
-		TIDs:     tids,
+		FirstTS:      a.firstTS,
+		LastTS:       a.lastTS,
+		Fingerprints: fingerprints,
 	}
 }
 
 // MetricsFileStats contains statistics about a metrics file.
 type MetricsFileStats struct {
-	TIDCount int32   // Number of unique TIDs in the file
-	MinTID   int64   // Minimum TID value
-	MaxTID   int64   // Maximum TID value
-	TIDs     []int64 // Actual list of unique TIDs in this file
+	FirstTS      int64   // Earliest timestamp
+	LastTS       int64   // Latest timestamp
+	Fingerprints []int64 // Fingerprints for indexing
 }
 
 // ValidateMetricsRow checks that a row has the required fields for metrics processing.
 func ValidateMetricsRow(row map[string]any) error {
+	if _, ok := row["_cardinalhq.name"]; !ok {
+		return fmt.Errorf("missing required field: _cardinalhq.name")
+	}
 	if _, ok := row["_cardinalhq.tid"]; !ok {
 		return fmt.Errorf("missing required field: _cardinalhq.tid")
 	}
